fix: getUserVote choice return type uint8 -> uint256 in frontend ABI
Matches the redeployed registry which forwards the Stylus engine's
uint256 choice value.

+++ frontend/src/lib/contracts.ts
@@ -77,7 +77,7 @@ export const REGISTRY_ABI = [
     ],
     name: "getUserVote",
     outputs: [
-      { name: "choice", type: "uint8" },
+      { name: "choice", type: "uint256" },
       { name: "weight", type: "uint256" },
     ],
     stateMutability: "view",
